+++ libgpopt/src/base/CColRefComputed.cpp
@@ -24,7 +24,7 @@ using namespace gpopt;
 //		takes ownership of string; verify string is properly formatted
 //
 //---------------------------------------------------------------------------
-CColRefComputed::CColRefComputed(const IMDType *pmdtype, INT type_modifier, ULONG id, const CName *pname)
+CColRefComputed::CColRefComputed(const IMDType *pmdtype, int32_t type_modifier, uint32_t id, const CName *pname)
     : CColRef(pmdtype, type_modifier, id, pname) {
   GPOS_ASSERT(nullptr != pmdtype);
   GPOS_ASSERT(pmdtype->MDId()->IsValid());
